LPS-155210 Refactor module structure
