+++ modules/apps/object/object-js-components-web/.eslintignore
@@ -1,0 +1,1 @@
+/types
