+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/js/components/Card.scss
