+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/js/global.d.ts
@@ -12,24 +12,8 @@
  * details.
  */
 
-type Locale =
-	| 'ar_SA'
-	| 'ca_ES'
-	| 'de_DE'
-	| 'en_US'
-	| 'es_ES'
-	| 'fi_FI'
-	| 'fr_FR'
-	| 'hu_HU'
-	| 'nl_NL'
-	| 'ja_JP'
-	| 'pt_BR'
-	| 'sv_SE'
-	| 'zh_CN';
-
-type LocalizedValue<T> = {
-	[key in Locale]?: T;
-};
+type Locale = Liferay.Language.Locale;
+type LocalizedValue<T> = Liferay.Language.LocalizedValue<T>;
 
 interface ObjectValidationRuleElement {
 	items: ObjectValidationRuleElementItem[];
