--- modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/js/index.d.ts
@@ -1,33 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-export {Input} from './core/components/Input';
-export {FieldBase} from './core/components/FieldBase';
-export {Card} from './core/components/Card/Card';
-export {InputLocalized} from './core/components/InputLocalized';
-export {CustomSelect} from './core/components/CustomSelect';
-export {FormCustomSelect, CustomItem} from './core/components/FormCustomSelect';
-export {AutoComplete} from './core/components/AutoComplete';
-export {CheckboxItem} from './core/components/CheckBoxItem';
-export {Select} from './core/components/Select';
-export {RichTextLocalized} from './core/components/RichTextLocalized/RichTextLocalized';
-export {Collapsible} from './core/components/Collapsible';
-export {Element} from './core/components/Element';
-export {EditorSidebar} from './core/components/EditorSideBar';
-export {
-	SidePanelContent,
-	SidePanelForm,
-	closeSidePanel,
-	openToast,
-} from './core/components/SidePanelContent';
