+++ modules/apps/object/object-js-components-web/types/src/main/resources/META-INF/resources/js/components/Select.d.ts
